+++ lib/typedef.h
@@ -3,7 +3,7 @@ using namespace std;
 typedef long long int i64;
 typedef unsigned long long int u64;
 typedef pair<int, int> IntPair;
-typedef pair<i64, int> I64Pair;
+typedef pair<i64, i64> I64Pair;
 typedef vector<int> IntVec ;
 typedef vector<IntPair> IntPairVec;
 typedef vector<I64Pair> I64PairVec;
